Fix gosec G115, slim SECURITY.md, move security docs to zatrano.com.

+++ bloom/bloom.go
@@ -102,8 +102,8 @@ func (f *Filter) indexes(key string) []uint64 {
 		h2 = 0x9e3779b97f4a7c15
 	}
 	out := make([]uint64, f.k)
-	for i := 0; i < f.k; i++ {
-		out[i] = (h1 + uint64(i)*h2) % f.m
+	for i := uint64(0); i < uint64(f.k); i++ {
+		out[i] = (h1 + i*h2) % f.m
 	}
 	return out
 }
